requirements.txt
Flask
werkzeug
pypdf
Pillow
rembg
img2pdf
pdf2image
reportlab
openpyxl
python-docx
pikepdf
python-pptx
xhtml2pdf
reportlab
Pillow
pdfminer.six
requests
beautifulsoup4
lxml
PyMuPDF
pdfplumber
tabula-py
camelot-py[cv]
numpy
pandas